pyproject.toml
[tool.poetry]
name = "top500-aggregator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "top500_aggregator"}]

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "4.5.5.64"
pillow = "^9.4.0"
pygetwindow = "^0.0.9"
pyautogui = "^0.9.53"
pytesseract = "^0.3.10"
rich = "^13.3.1"
fastapi = "^0.89.1"
jinja2 = "^3.1.2"
uvicorn = "^0.20.0"
matplotlib = "^3.6.3"
nnfs = "^0.5.1"
numpy = "^1.24.3"
pytest = "^7.3.1"
hypercorn = "^0.14.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
